fix: PR #18 review responses + version bump to 1.0.2
Review fixes (Codacy / CodeRabbit):
- error_handling.ino: cast days_since_epoch to uint32_t before multiply (signed overflow UB)
- rtos_freertos.ino: correct borrowed-pointer comment (toJson does NOT copy internally)
- avr_std/optional: operator-> uses storage_ directly, bypasses overloaded operator&
- TimeFields.h: accurate worst-case description (non-monotonic, epoch loss)
- test/Makefile: single-shell memprof recipe, skip guard now functional
- test_avr_std_shim.cpp: regression test for operator& overload bypass

Version sync:
- library.json: 1.0.0 → 1.0.2
- library.properties: 1.0.1 → 1.0.2
- CHANGELOG.md: add 1.0.2 entry

+++ examples/error_handling/error_handling.ino
@@ -71,10 +71,10 @@ static uint32_t to_unix_time(uint16_t year, uint8_t month, uint8_t day,
     uint32_t doy = (153 * (m + (m > 2 ? -3 : 9)) + 2) / 5 + d - 1;
     uint32_t doe = yoe * 365 + yoe / 4 - yoe / 100 + doy;
     int32_t days_since_epoch = (int32_t)(era * 146097 + doe - 719468);
-    return (uint32_t)((days_since_epoch * 86400L) +
-                      (hour * 3600L) +
-                      (minute * 60L) +
-                      sec);
+    return (uint32_t)((uint32_t)days_since_epoch * 86400UL +
+                      (uint32_t)hour * 3600UL +
+                      (uint32_t)minute * 60UL +
+                      (uint32_t)sec);
 }
 
 static uint32_t cached_gnss_unix_time = 0;
